requirements.txt
absl-py==0.15.0
aniso8601==9.0.1
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
EasyProcess==0.3
Flask==2.0.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.5.1
flatbuffers==2.0
gast==0.4.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.41.1
h11==0.12.0
h5py==3.5.0
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
keras==2.7.0
Keras-Preprocessing==1.1.2
libclang==12.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
outcome==1.1.0
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==21.0.0
pytz==2021.3
PyVirtualDisplay==2.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.26
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.3.0
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wget==3.2
wrapt==1.13.3
wsproto==1.0.0
